refactor: move layout_text type implementation into new text folder
